requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
opencv-python==4.5.5.62
Pillow==9.5.0
google-cloud-storage==2.9.0
boto3==1.28.68
tqdm==4.65.0